404, edição de arquivos em desenvolvimento, fixes. css do context menu.

+++ src/controllers/Controller.php
@@ -1,8 +1,11 @@
 <?php
+require_once __DIR__ . "/../models/Singleton.php";
 require_once __DIR__ . "/IndexController.php";
 require_once __DIR__ . "/AccessController.php";
 require_once __DIR__ . "/LogoutController.php";
-require_once __DIR__ . "/../models/Singleton.php";
+require_once __DIR__ . "/EditController.php";
+require_once __DIR__ . "/ErrorController.php";
+require_once __DIR__ . "/../helpers/info.php";
 
 session_start();
 
@@ -11,6 +14,8 @@ class Controller {
     private $access;
     private $index;
     private $logout;
+    private $edit;
+    private $error;
 
     public function __construct() {
         $this->mysqli = new Singleton();
@@ -18,6 +23,8 @@ class Controller {
         $this->access = new AccessController($this->mysqli);
         $this->index = new IndexController($this->mysqli);
         $this->logout = new LogoutController();
+        $this->edit = new EditController();
+        $this->error = new ErrorController();
     }
 
     public function render__access() {
@@ -30,16 +37,20 @@ class Controller {
         if (isset($_GET["id"])) :
             if (isset($_GET["edit"]) || isset($_GET["delete"])):
                 // EXCLUSÃO DE ARQUIVO
-                if (isset($_GET["delete"])):
+                if (isset($_GET["delete"]) && !(isset($_GET["edit"]))):
                     $sql = "DELETE FROM files WHERE id = " . $_GET['id'];
                     $query = $this->mysqli->getInstance()->query($sql);
 
                     if (!$query) {
-                        echo "Erro.";
+                        echo info__show("Erro.", "bg-danger");
                     }
                     $this->mysqli->getInstance()->commit();
                 endif;
 
+                // if (isset($_GET["edit"]) && !(isset($_GET["delete"]))):
+                //     header("location: /edit?" . $_SERVER['QUERY_STRING']);
+                //     exit;
+                // endif;
             // DOWNLOAD DO ARQUIVO
             else:
                 $sql = "SELECT * FROM files WHERE id = " . $_GET['id'];
@@ -63,12 +74,24 @@ class Controller {
             endif;
         endif;
 
-        $this->mysqli->unsetInstance();
-        $this->index->load();
+        if (!(isset($_GET["edit"]))):
+            $this->index->load();
+            $this->mysqli->unsetInstance();
+        endif;
     }
 
     public function render__logout() {
         $this->logout->load();
+    }
+
+    public function render__edit() {
+        $this->edit->load();
+    }
+
+    public function render__error($error) {
+        if ($error == "404"):
+            $this->error->load__404();
+        endif;
     }
 }
 ?>
